+++ src/mo_config/flags.ml
@@ -72,7 +72,7 @@ let use_stable_regions = ref false
 let enhanced_orthogonal_persistence = ref true
 let explicit_enhanced_orthogonal_persistence = ref false
 let enhanced_migration : string option ref = ref None
-(* Last deployed .most for check-time stable-compat + M0254 carry *)
+(* Last deployed .most for the resume-point boundary check *)
 let stable_baseline : string option ref = ref None
 let share_code = ref false
 let stabilization_instruction_limit_default = {